refactor(webhooks): Wire legacy webhook service directly from action handler
Move feature-flagged routing from WebhookIssueAlertHandler.invoke_legacy_registry
up to WebhookActionHandler.execute so the new path calls
send_legacy_webhooks_for_invocation directly, bypassing the NOA registry
dispatch chain. The old path still falls through to
execute_via_group_type_registry when not disabled.
