+++ src/sentry/notifications/notification_action/action_handler_registry/webhook_handler.py
@@ -1,6 +1,8 @@
 from typing import override
 
+from sentry import features
 from sentry.notifications.notification_action.utils import execute_via_group_type_registry
+from sentry.sentry_apps.services.legacy_webhook.service import send_legacy_webhooks_for_invocation
 from sentry.workflow_engine.models import Action
 from sentry.workflow_engine.registry import action_handler_registry
 from sentry.workflow_engine.types import ActionHandler, ActionInvocation, ConfigTransformer
@@ -36,4 +38,12 @@ class WebhookActionHandler(ActionHandler):
     @staticmethod
     @override
     def execute(invocation: ActionInvocation) -> None:
-        execute_via_group_type_registry(invocation)
+        organization = invocation.detector.project.organization
+        new_path = features.has("organizations:legacy-webhook-new-path", organization)
+        disable_old = features.has("organizations:legacy-webhook-disable-old-path", organization)
+
+        if new_path:
+            send_legacy_webhooks_for_invocation(invocation)
+
+        if not disable_old:
+            execute_via_group_type_registry(invocation)
