+++ tests/sentry/notifications/notification_action/test_issue_alert_registry_handlers.py
@@ -2,15 +2,10 @@ import uuid
 from unittest import mock
 
 import pytest
-import responses
 
 from sentry.constants import ObjectStatus
-from sentry.models.options.project_option import ProjectOption
 from sentry.models.rule import Rule, RuleSource
 from sentry.notifications.models.notificationaction import ActionTarget
-from sentry.notifications.notification_action.group_type_notification_registry.handlers.issue_alert_registry_handler import (
-    IssueAlertRegistryHandler,
-)
 from sentry.notifications.notification_action.issue_alert_registry import (
     AzureDevopsIssueAlertHandler,
     DiscordIssueAlertHandler,
@@ -31,7 +26,6 @@ from sentry.notifications.notification_action.types import (
     TicketingIssueAlertHandler,
 )
 from sentry.notifications.types import TEST_NOTIFICATION_ID, ActionTargetType, FallthroughChoiceType
-from sentry.plugins.base import plugins
 from sentry.testutils.helpers.data_blobs import (
     AZURE_DEVOPS_ACTION_DATA_BLOBS,
     EMAIL_ACTION_DATA_BLOBS,
@@ -40,7 +34,6 @@ from sentry.testutils.helpers.data_blobs import (
     JIRA_SERVER_ACTION_DATA_BLOBS,
     WEBHOOK_ACTION_DATA_BLOBS,
 )
-from sentry.utils import json
 from sentry.workflow_engine.models import Action
 from sentry.workflow_engine.types import ActionInvocation, WorkflowEventData
 from sentry.workflow_engine.typings.notification_action import (
@@ -672,98 +665,6 @@ class TestWebhookIssueAlertHandler(BaseWorkflowTest):
             blob = self.handler.build_rule_action_blob(action, self.organization.id)
 
             assert blob == expected
-
-
-class TestWebhookIssueAlertHandlerInvokeLegacyRegistry(BaseWorkflowTest):
-    def setUp(self) -> None:
-        super().setUp()
-        self.detector = self.create_detector(project=self.project)
-        self.workflow = self.create_workflow(environment=self.environment)
-        self.action = self.create_action(
-            type=Action.Type.WEBHOOK,
-            config={"target_identifier": "webhooks"},
-        )
-        self.group, self.event, self.group_event = self.create_group_event()
-        self.event_data = WorkflowEventData(
-            event=self.group_event, workflow_env=self.environment, group=self.group
-        )
-        self.invocation = ActionInvocation(
-            event_data=self.event_data,
-            action=self.action,
-            detector=self.detector,
-            notification_uuid=str(uuid.uuid4()),
-            workflow_id=self.workflow.id,
-        )
-        ProjectOption.objects.set_value(self.project, "webhooks:urls", "http://example.com/hook")
-        webhook_plugin = plugins.get("webhooks")
-        webhook_plugin.set_option("enabled", True, self.project)
-
-    @responses.activate
-    def test_default_no_flags_fires_old_path_only(self) -> None:
-        responses.add(responses.POST, "http://example.com/hook")
-
-        with self.tasks():
-            IssueAlertRegistryHandler.handle_workflow_action(self.invocation)
-
-        assert len(responses.calls) == 1
-
-    @responses.activate
-    def test_new_path_fires_both_paths(self) -> None:
-        responses.add(responses.POST, "http://example.com/hook")
-
-        with self.tasks(), self.feature("organizations:legacy-webhook-new-path"):
-            IssueAlertRegistryHandler.handle_workflow_action(self.invocation)
-
-        assert len(responses.calls) == 2
-
-    @responses.activate
-    def test_new_path_disable_old_fires_new_only(self) -> None:
-        responses.add(responses.POST, "http://example.com/hook")
-
-        with (
-            self.tasks(),
-            self.feature(
-                {
-                    "organizations:legacy-webhook-new-path": True,
-                    "organizations:legacy-webhook-disable-old-path": True,
-                }
-            ),
-        ):
-            IssueAlertRegistryHandler.handle_workflow_action(self.invocation)
-
-        assert len(responses.calls) == 1
-        body = json.loads(responses.calls[0].request.body)
-        assert body["id"] == str(self.group.id)
-
-    @responses.activate
-    @mock.patch("sentry.sentry_apps.services.legacy_webhook.tasks.logger")
-    def test_new_path_dry_run_logs_instead_of_sending(self, mock_logger: mock.MagicMock) -> None:
-        responses.add(responses.POST, "http://example.com/hook")
-
-        with (
-            self.tasks(),
-            self.feature(
-                {
-                    "organizations:legacy-webhook-new-path": True,
-                    "organizations:legacy-webhook-disable-old-path": True,
-                    "organizations:legacy-webhook-dry-run": True,
-                }
-            ),
-        ):
-            IssueAlertRegistryHandler.handle_workflow_action(self.invocation)
-
-        assert len(responses.calls) == 0
-        mock_logger.info.assert_called_once()
-        assert mock_logger.info.call_args[0][0] == "legacy_webhook.dry_run"
-
-    @responses.activate
-    def test_disable_old_without_new_path_fires_nothing(self) -> None:
-        responses.add(responses.POST, "http://example.com/hook")
-
-        with self.tasks(), self.feature("organizations:legacy-webhook-disable-old-path"):
-            IssueAlertRegistryHandler.handle_workflow_action(self.invocation)
-
-        assert len(responses.calls) == 0
 
 
 class TestSentryAppIssueAlertHandler(BaseWorkflowTest):
